+++ Controls/_baseList/ListView.ts
@@ -8,8 +8,7 @@ import * as GroupTemplate from 'wml!Controls/_baseList/GroupTemplate';
 import * as ListViewTpl from 'wml!Controls/_baseList/ListView/ListView';
 import * as defaultItemTemplate from 'wml!Controls/_baseList/ItemTemplate';
 import 'css!Controls/baseList';
-import {Collection} from 'Controls/display';
-import {IRoundBorder} from 'Controls/interface';
+import {Collection, IRoundBorder} from 'Controls/display';
 
 export interface IListViewOptions {
     listModel: Collection;
