+++ Controls/_display/Collection.ts
@@ -35,7 +35,8 @@ import {Object as EventObject} from 'Env/Event';
 import * as VirtualScrollController from './controllers/VirtualScroll';
 import { ICollection, ISourceCollection, IItemPadding } from './interface/ICollection';
 import { IDragPosition } from './interface/IDragPosition';
-import {INavigationOptionValue, INavigationSourceConfig, IRoundBorder} from 'Controls/interface';
+import { IRoundBorder } from './interface/IRoundBorder';
+import {INavigationOptionValue, INavigationSourceConfig} from 'Controls/interface';
 import {Footer} from 'Controls/_display/Footer';
 import IndicatorsMixin from './IndicatorsMixin';
 import {Logger} from 'UI/Utils';
