+++ Controls/_display/interface/ICollection.ts
@@ -2,7 +2,6 @@ import { DestroyableMixin, Model, ObservableMixin } from 'Types/entity';
 import {IEnumerable} from '../Abstract';
 import {IBaseCollection} from 'Controls/_display/interface';
 import {ICollectionItem} from 'Controls/_display/interface/ICollectionItem';
-import {IRoundBorder} from 'Controls/interface';
 
 export interface ISourceCollection<T extends Model = Model> extends IEnumerable<T>, DestroyableMixin, ObservableMixin {
     getCount(): number;
@@ -97,8 +96,6 @@ export enum ANIMATION_STATE {
  * @private
  * @author Аверкиев П.А.
  */
-
-export type TRoundBorder = IRoundBorder;
 
 export interface ICollection<S extends Model, T extends ICollectionItem> extends IBaseCollection<S, T> {
     getCollection(): ISourceCollection<S>;
