+++ Controls/_tile/display/mixins/Tile.ts
@@ -2,7 +2,7 @@ import {Model} from 'Types/entity';
 import TileItem, { IOptions as ITileItemOptions } from 'Controls/_tile/display/mixins/TileItem';
 import {isEqual} from 'Types/object';
 import {IViewIterator, IItemPadding, IItemActionsTemplateConfig} from 'Controls/display';
-import {IRoundBorder} from 'Controls/interface';
+import {ITileRoundBorder} from 'Controls/interface';
 import {createPositionInBounds} from 'Controls/_tile/utils/createPosition';
 
 export const DEFAULT_TILE_HEIGHT = 200;
@@ -56,7 +56,7 @@ export default abstract class Tile<
 
     protected _$itemsContainerPadding: IItemPadding;
 
-    protected _$roundBorder: IRoundBorder;
+    protected _$roundBorder: ITileRoundBorder;
 
     protected _$imageProperty: string;
 
@@ -358,18 +358,18 @@ export default abstract class Tile<
 
     /**
      * Возвращает скругление углов элемента
-     * @return {IRoundBorder} Cкругление углов элемента плитки
+     * @return {ITileRoundBorder} Cкругление углов элемента плитки
      */
-    getRoundBorder(): IRoundBorder {
+    getRoundBorder(): ITileRoundBorder {
         return this._$roundBorder;
     }
 
     /**
      * Устанавливает скругление углов элемента
-     * @param {IRoundBorder} roundBorder Скругление углов элемента плитки
+     * @param {ITileRoundBorder} roundBorder Скругление углов элемента плитки
      * @void
      */
-    setRoundBorder(roundBorder: IRoundBorder): void {
+    setRoundBorder(roundBorder: ITileRoundBorder): void {
         if (!isEqual(this._$roundBorder, roundBorder)) {
             this._$roundBorder = roundBorder;
             this._updateItemsProperty(
