+++ Controls/_tile/display/mixins/TileItem.ts
@@ -17,7 +17,7 @@ import Tile, {
     TImageUrlResolver, TTileMode, TTileScalingMode, TTileSize
 } from './Tile';
 import { TItemActionsPosition } from 'Controls/itemActions';
-import {IRoundBorder} from 'Controls/interface';
+import {ITileRoundBorder} from 'Controls/interface';
 import {TBackgroundColorStyle, TCursor } from 'Controls/list';
 import {toRgb, rgbaToString, rgbToRgba} from 'Controls/Utils/colorUtil';
 
@@ -79,7 +79,7 @@ export interface IOptions<S extends Model = Model> extends ICollectionItemOption
     tileHeight: number;
     tileWidth: number;
     tileWidthProperty: string;
-    roundBorder: IRoundBorder;
+    roundBorder: ITileRoundBorder;
     imageProperty: string;
     imageFit: TImageFit;
     imageHeightProperty: string;
@@ -104,7 +104,7 @@ export default abstract class TileItem<T extends Model = Model> {
 
     protected _$canShowActions: boolean;
 
-    protected _$roundBorder: IRoundBorder;
+    protected _$roundBorder: ITileRoundBorder;
 
     // region TileOptions
 
@@ -1861,10 +1861,10 @@ export default abstract class TileItem<T extends Model = Model> {
 
     /**
      * Устанавливает скругление углов элемента
-     * @param {IRoundBorder} roundBorder Скругление углов элемента
+     * @param {ITileRoundBorder} roundBorder Скругление углов элемента
      * @void
      */
-    setRoundBorder(roundBorder: IRoundBorder): void {
+    setRoundBorder(roundBorder: ITileRoundBorder): void {
         if (!isEqual(this._$roundBorder, roundBorder)) {
             this._$roundBorder = roundBorder;
             this._nextVersion();
