+++ Controls/display.ts
@@ -65,10 +65,11 @@ export {BreadcrumbsItem};
 import SearchSeparator from './_display/SearchSeparator';
 export {SearchSeparator};
 
-export {ANIMATION_STATE, IItemPadding, TRoundBorder} from './_display/interface/ICollection';
+export {ANIMATION_STATE, IItemPadding} from './_display/interface/ICollection';
 export {IEditableCollection} from './_display/interface/IEditableCollection';
 export {IEditableCollectionItem} from './_display/interface/IEditableCollectionItem';
 export {ICollectionItem} from './_display/interface/ICollectionItem';
+export {IRoundBorder} from './_display/interface/IRoundBorder';
 export {IBaseCollection, TItemKey} from './_display/interface';
 
 import * as GridLadderUtil from './_display/utils/GridLadderUtil';
